Fix file save. First round trip save-to-treemap works ------------
- quoted 7z paths: spaced names
- release mode / publish
- preserve start and end times on convert
- start drive chooser for drive model/serial
- MainWindow readonly internal var; replaces Window private set prop

+++ DoubleFile/DoubleFile/UtilAnalysis_DirList/Tree/TreeBase/Tree/TreeRootNodeBuilder.cs
@@ -229,8 +229,8 @@ namespace DoubleFile
                 if (bZeroLengthsWritten)
                 {
 #if (DEBUG)
-                    UtilProject.WriteLine(File.ReadLines(m_volStrings.ListingFile).Where(s => s.StartsWith(ksLineType_File)).Sum(s => decimal.Parse(s.Split('\t')[knColLength])).ToString());
-                    UtilProject.WriteLine(File.ReadLines(m_volStrings.ListingFile).Where(s => s.StartsWith(ksLineType_Directory)).Sum(s => decimal.Parse(s.Split('\t')[knColLength])).ToString());
+                    UtilProject.WriteLine(File.ReadLines(m_volStrings.ListingFile).Where(s => s.StartsWith(ksLineType_File)).Sum(s => double.Parse(s.Split('\t')[knColLength])).ToString());
+                    UtilProject.WriteLine(File.ReadLines(m_volStrings.ListingFile).Where(s => s.StartsWith(ksLineType_Directory)).Sum(s => double.Parse(s.Split('\t')[knColLength])).ToString());
 #endif
                 }
